MSVC fixes
* Removed alignas, makes things even worse on MSVC, which is now slower on my pimped out 2020 DOOM slayer than my MBP 2016 :(
* Changed io::timeMillis() to return a double instead of an uint64_t.
* Fixed all warnings with MSVC 2019.

+++ src/apps/test_parser.cpp
@@ -13,7 +13,7 @@ void testBench() {
     Source *source = io::readFile("data/parser_function.qak", mem);
     QAK_CHECK(source != nullptr, "Couldn't read test file data/parser_benchmark.qak");
 
-    uint64_t start = io::timeMillis();
+    double start = io::timeMillis();
     Parser parser(mem);
     Errors errors(mem);
