+++ src/apps/test_tokenizer.cpp
@@ -7,7 +7,7 @@ using namespace qak;
 
 void testBench() {
     Test test("Tokenizer - Benchmark");
-    uint64_t start = io::timeMillis();
+    double start = io::timeMillis();
     HeapAllocator mem;
     Source *source = io::readFile("data/tokens.qak", mem);
     QAK_CHECK(source != nullptr, "Couldn't read test file data/tokens.qak");
